refactor: Re-organize the layout of tests (#3904)
* Refactor tests


* refactor: Re-organize the layout of tests


---------
