+++ core/tests/behavior/blocking_stat.rs
@@ -1,0 +1,107 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+use anyhow::Result;
+use log::debug;
+use log::warn;
+
+use crate::*;
+
+pub fn tests(op: &Operator, tests: &mut Vec<Trial>) {
+    let cap = op.info().full_capability();
+
+    if cap.stat && cap.write && cap.blocking {
+        tests.extend(blocking_trials!(
+            op,
+            test_blocking_stat_file,
+            test_blocking_stat_dir,
+            test_blocking_stat_with_special_chars,
+            test_blocking_stat_not_exist
+        ))
+    }
+}
+
+/// Stat existing file should return metadata
+pub fn test_blocking_stat_file(op: BlockingOperator) -> Result<()> {
+    let path = uuid::Uuid::new_v4().to_string();
+    debug!("Generate a random file: {}", &path);
+    let (content, size) = gen_bytes(op.info().full_capability());
+
+    op.write(&path, content).expect("write must succeed");
+
+    let meta = op.stat(&path)?;
+    assert_eq!(meta.mode(), EntryMode::FILE);
+    assert_eq!(meta.content_length(), size as u64);
+
+    op.delete(&path).expect("delete must succeed");
+    Ok(())
+}
+
+/// Stat existing file should return metadata
+pub fn test_blocking_stat_dir(op: BlockingOperator) -> Result<()> {
+    if !op.info().full_capability().create_dir {
+        return Ok(());
+    }
+
+    let path = format!("{}/", uuid::Uuid::new_v4());
+
+    op.create_dir(&path).expect("write must succeed");
+
+    let meta = op.stat(&path)?;
+    assert_eq!(meta.mode(), EntryMode::DIR);
+
+    op.delete(&path).expect("delete must succeed");
+    Ok(())
+}
+
+/// Stat existing file with special chars should return metadata
+pub fn test_blocking_stat_with_special_chars(op: BlockingOperator) -> Result<()> {
+    // Ignore test for supabase until https://github.com/apache/incubator-opendal/issues/2194 addressed.
+    if op.info().scheme() == opendal::Scheme::Supabase {
+        warn!("ignore test for supabase until https://github.com/apache/incubator-opendal/issues/2194 is resolved");
+        return Ok(());
+    }
+    // Ignore test for atomicserver until https://github.com/atomicdata-dev/atomic-server/issues/663 addressed.
+    if op.info().scheme() == opendal::Scheme::Atomicserver {
+        warn!("ignore test for atomicserver until https://github.com/atomicdata-dev/atomic-server/issues/663 is resolved");
+        return Ok(());
+    }
+
+    let path = format!("{} !@#$%^&()_+-=;',.txt", uuid::Uuid::new_v4());
+    debug!("Generate a random file: {}", &path);
+    let (content, size) = gen_bytes(op.info().full_capability());
+
+    op.write(&path, content).expect("write must succeed");
+
+    let meta = op.stat(&path)?;
+    assert_eq!(meta.mode(), EntryMode::FILE);
+    assert_eq!(meta.content_length(), size as u64);
+
+    op.delete(&path).expect("delete must succeed");
+    Ok(())
+}
+
+/// Stat not exist file should return NotFound
+pub fn test_blocking_stat_not_exist(op: BlockingOperator) -> Result<()> {
+    let path = uuid::Uuid::new_v4().to_string();
+
+    let meta = op.stat(&path);
+    assert!(meta.is_err());
+    assert_eq!(meta.unwrap_err().kind(), ErrorKind::NotFound);
+
+    Ok(())
+}
