+++ core/tests/behavior/main.rs
@@ -21,44 +21,34 @@ mod utils;
 
 pub use utils::*;
 
-// Async test cases
-mod append;
-mod copy;
-mod fuzz;
-mod list;
-mod list_only;
-mod presign;
-mod read_only;
-mod rename;
-mod write;
-use append::behavior_append_tests;
-use copy::behavior_copy_tests;
-use fuzz::behavior_fuzz_tests;
-use list::behavior_list_tests;
-use list_only::behavior_list_only_tests;
-use presign::behavior_presign_tests;
-use read_only::behavior_read_only_tests;
-use rename::behavior_rename_tests;
-use write::behavior_write_tests;
+mod async_copy;
+mod async_create_dir;
+mod async_delete;
+mod async_fuzz;
+mod async_list;
+mod async_presign;
+mod async_read;
+mod async_rename;
+mod async_stat;
+mod async_write;
 
 // Blocking test cases
-mod blocking_append;
 mod blocking_copy;
+mod blocking_create_dir;
+mod blocking_delete;
 mod blocking_list;
-mod blocking_read_only;
+mod blocking_read;
 mod blocking_rename;
+mod blocking_stat;
 mod blocking_write;
-use blocking_append::behavior_blocking_append_tests;
-use blocking_copy::behavior_blocking_copy_tests;
-use blocking_list::behavior_blocking_list_tests;
-use blocking_read_only::behavior_blocking_read_only_tests;
-use blocking_rename::behavior_blocking_rename_tests;
-use blocking_write::behavior_blocking_write_tests;
+
 // External dependencies
 use libtest_mimic::Arguments;
 use libtest_mimic::Trial;
-use opendal::raw::tests::init_test_service;
+use opendal::raw::tests::{init_test_service, TEST_RUNTIME};
 use opendal::*;
+
+pub static TEST_FIXTURE: Fixture = Fixture::new();
 
 fn main() -> anyhow::Result<()> {
     let args = Arguments::from_args();
@@ -70,23 +60,26 @@ fn main() -> anyhow::Result<()> {
     };
 
     let mut tests = Vec::new();
-    // Blocking tests
-    tests.extend(behavior_blocking_append_tests(&op));
-    tests.extend(behavior_blocking_copy_tests(&op));
-    tests.extend(behavior_blocking_list_tests(&op));
-    tests.extend(behavior_blocking_read_only_tests(&op));
-    tests.extend(behavior_blocking_rename_tests(&op));
-    tests.extend(behavior_blocking_write_tests(&op));
-    // Async tests
-    tests.extend(behavior_append_tests(&op));
-    tests.extend(behavior_copy_tests(&op));
-    tests.extend(behavior_list_only_tests(&op));
-    tests.extend(behavior_list_tests(&op));
-    tests.extend(behavior_presign_tests(&op));
-    tests.extend(behavior_read_only_tests(&op));
-    tests.extend(behavior_rename_tests(&op));
-    tests.extend(behavior_write_tests(&op));
-    tests.extend(behavior_fuzz_tests(&op));
+
+    async_copy::tests(&op, &mut tests);
+    async_create_dir::tests(&op, &mut tests);
+    async_delete::tests(&op, &mut tests);
+    async_fuzz::tests(&op, &mut tests);
+    async_list::tests(&op, &mut tests);
+    async_presign::tests(&op, &mut tests);
+    async_read::tests(&op, &mut tests);
+    async_rename::tests(&op, &mut tests);
+    async_stat::tests(&op, &mut tests);
+    async_write::tests(&op, &mut tests);
+
+    blocking_copy::tests(&op, &mut tests);
+    blocking_create_dir::tests(&op, &mut tests);
+    blocking_delete::tests(&op, &mut tests);
+    blocking_list::tests(&op, &mut tests);
+    blocking_read::tests(&op, &mut tests);
+    blocking_rename::tests(&op, &mut tests);
+    blocking_stat::tests(&op, &mut tests);
+    blocking_write::tests(&op, &mut tests);
 
     // Don't init logging while building operator which may break cargo
     // nextest output
@@ -96,5 +89,10 @@ fn main() -> anyhow::Result<()> {
         .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
         .try_init();
 
-    libtest_mimic::run(&args, tests).exit();
+    let conclusion = libtest_mimic::run(&args, tests);
+
+    // Cleanup the fixtures.
+    TEST_RUNTIME.block_on(TEST_FIXTURE.cleanup(op));
+
+    conclusion.exit()
 }
